+++ pcs_test/tools/command_env/config_runner_pcmk.py
@@ -695,6 +695,7 @@ class PcmkShortcuts:
         stdout="",
         stderr="",
         returncode=0,
+        env=None,
     ):
         """
         Create a call for crm_resource --clear
@@ -711,6 +712,7 @@ class PcmkShortcuts:
         string stdout -- crm_resource's stdout
         string stderr -- crm_resource's stderr
         int returncode -- crm_resource's returncode
+        dict env -- CommandRunner environment variables
         """
         # arguments are used via locals()
         # pylint: disable=unused-argument